GH-44: add smarter custom parameters handler
It wasn't possible to get custom class-level parameters due to the fact we get only the first config that matches the browser. However, it didn't mean it had custom params. This update contains a new logic of extracting all the custom params beforehand with further merging with the browser config.